frontend package now exports a build function
